Functions for getting features now started to splitting into separated module.
