+++ src/codeplag/types.py
@@ -1,4 +1,6 @@
-from typing import Dict, List, NamedTuple, TypedDict
+from collections import defaultdict
+from pathlib import Path
+from typing import Dict, List, Literal, NamedTuple, TypedDict
 
 import numpy as np
 
@@ -11,6 +13,26 @@ class NodeCodePlace(NamedTuple):
 class NodeStructurePlace(NamedTuple):
     depth: int
     uid: int
+
+
+class ASTFeatures:
+    def __init__(self, filepath: Path):
+        self.filepath = filepath
+
+        self.count_of_nodes = 0
+        self.head_nodes: List[str] = []
+        self.operators: Dict[str, int] = defaultdict(lambda: 0)
+        self.keywords: Dict[str, int] = defaultdict(lambda: 0)
+        self.literals: Dict[str, int] = defaultdict(lambda: 0)
+
+        # unique nodes
+        self.unodes: Dict[str, int] = {}
+        self.from_num: Dict[int, str] = {}
+        self.count_unodes = 0
+
+        self.structure: List[NodeStructurePlace] = []
+        self.tokens: List[int] = []
+        self.tokens_pos: List[NodeCodePlace] = []
 
 
 class FastMetrics(NamedTuple):
@@ -39,3 +61,6 @@ class WorksReport(TypedDict):
     second_heads: List[str]
     fast: Dict[str, int]
     structure: dict
+
+
+Mode = Literal["many_to_many", "one_to_one"]
