+++ test/unit/codeplag/cplag/test_tree.py
@@ -4,10 +4,10 @@ from pathlib import Path
 
 from clang.cindex import CursorKind
 
-from codeplag.astfeatures import ASTFeatures
 from codeplag.cplag.const import COMPILE_ARGS
 from codeplag.cplag.tree import generic_visit, get_features, get_not_ignored
 from codeplag.cplag.util import get_cursor_from_file
+from codeplag.types import ASTFeatures
 
 
 class TestTree(unittest.TestCase):
